Refactored test_kinematics.py similar to kinematics.py.
